+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.8.0] - 2026-02-26
 ### Fixed
 - Main bgm not correctly unmuted after enemy cleared
+- Note priority, which caused N64 stealing soundtrack notes whenever it felt like it
 ### Changed
 - better concern separation to make it easier to control audio behavior without locking down API to a specific use case